@@ -1,57 +1,7 @@
-###
-# Compass
-###
-
-# Susy grids in Compass
-# First: gem install susy
-# require 'susy'
-
-# Change Compass configuration
-# compass_config do |config|
-#   config.output_style = :compact
-# end
-
-###
-# Page options, layouts, aliases and proxies
-###
-
-# Per-page layout changes:
-#
-# With no layout
-# page "/path/to/file.html", :layout => false
 page "/feed.xml", layout: false
-#
-# With alternative layout
-# page "/path/to/file.html", :layout => :otherlayout
-#
-# A path which all have the same layout
-# with_layout :admin do
-#   page "/admin/*"
-# end
-with_layout :slides do
-  page "/slides/*"
-end
+page "/slides/*", layout: 'slides'
 
-set :layout, 'layouts/blog'
-
-# Proxy (fake) files
-# page "/this-page-has-no-template.html", :proxy => "/template-file.html" do
-#   @which_fake_page = "Rendering a fake page with a variable"
-# end
-
-###
-# Helpers
-###
-
-# Automatic image dimensions on image_tag helper
-# activate :automatic_image_sizes
-
-# Methods defined in the helpers block are available in templates
-# helpers do
-#   def some_helper
-#     "Helping"
-#   end
-# end
+set :layout, 'home'
 
 set :css_dir, 'stylesheets'
 
@@ -68,6 +18,7 @@ set :markdown, :fenced_code_blocks => true,
                :tables => true,
                :no_intra_emphasis => true
 
+set :relative_links, true
 activate :syntax
 
 activate :blog do |blog|
@@ -89,29 +40,9 @@ end
 activate :directory_indexes
 set :trailing_slash, false
 
-
-# Build-specific configuration
 configure :build do
-  # For example, change the Compass output style for deployment
   activate :minify_css
-
-  # Minify Javascript on build
   activate :minify_javascript
-
-  # Enable cache buster
-  #activate :cache_buster
-  # OR
-  # Enable asset hashes
   activate :asset_hash
-
-  # Use relative URLs
   activate :relative_assets
-
-  # Compress PNGs after build
-  # First: gem install middleman-smusher
-  # require "middleman-smusher"
-  # activate :smusher
-
-  # Or use a different image path
-  # set :http_path, "/Content/images/"
 end
